Simulate Weather Data Using MSTWeatherGen

This function simulates weather data over specified dates using the Multivariate Space-Time Stochastic Weather Generator (MSTWeatherGen). The simulation can be conducted on either a seasonal basis or for the entire period, depending on the provided parameters and the structure of the historical data. It utilizes autoregressive models and specified parameters to generate realistic weather variables.

MSTWeatherGen_Sim(
  dates_sim,
  dates_original,
  data,
  seasons = NULL,
  parm,
  AR_lag = 1,
  bk
)

Arguments

dates_sim

Dates for which to simulate weather data, specifying the target period for simulation.

dates_original

The original dates corresponding to the historical weather data, used for aligning and deriving simulation parameters.

data

Historical weather data array, serving as the basis for deriving parameters for simulation models and ensuring realistic weather patterns.

seasons

Optional definitions of seasons, which allows for applying different simulation parameters or models according to season-specific dynamics.

parm

A comprehensive parameters object containing essential data and model parameters for conducting the weather simulation.

AR_lag

The lag order of the autoregressive (AR) model used in the simulation, dictating how previous weather data influences future simulations.

bk

Coefficients matrices for the AR model, which may vary by season or weather type, playing a key role in the temporal modeling of weather variables.

Value

A 3D array (`sim`) of simulated weather data for the specified dates, encapsulating simulated values for each weather variable across spatial locations and over time.