Estimate Geostatistical Parameters for Multivariate Spatio-Temporal Data

This function estimates the geostatistical parameters for a multivariate space-time model based on observed spatio-temporal data. It integrates various preprocessing and optimization steps to derive the optimal model parameters that best fit the observed data.

estimation_gf(
  data,
  wt_id,
  max_it,
  dates,
  tmax,
  names,
  par_all = NULL,
  coordinates,
  n1,
  n2,
  ax,
  cr,
  threshold_precip
)

Arguments

data

A 3D array containing the observed data values across time, space, and variables.

wt_id

Vector of identifiers indicating specific weather types or variable categorizations within the dataset.

max_it

Maximum number of iterations allowed during the optimization process.

dates

Vector of dates corresponding to the temporal observations within the dataset.

tmax

The maximum temporal lag considered for the model, influencing temporal dependency estimation.

names

Vector containing the names of the variables included in the analysis.

par_all

(Optional) Initial or current complete set of model parameters. If not provided, parameters are initialized within the function.

coordinates

Matrix containing the geographical coordinates of the spatial locations in the dataset.

n1,

n2 Parameters that define the granularity for generating spatial index pairs, affecting the spatial resolution of the model.

ax

Matrix of precomputed correction terms used to adjust the covariance matrix, aiding in model stabilization.

cr

Initial correlation matrix representing the base relationships between variables, used as a starting point for optimization.

threshold_precip

Threshold values for precipitation, used in preprocessing to distinguish between different precipitation intensities.

Value

A list containing two elements: `parm`, the parameter matrix formatted for interpretation and use in subsequent model applications, and `par_all`, the complete vector of optimized model parameters.