Compute Log-likelihood for Variable Pair

Calculates the log-likelihood for a given pair of variables using the Gneiting spatio-temporal covariance model. This function is part of the internal mechanism for optimizing model parameters based on observed data.

loglik_pair(par, parms, pair, par_all, data, names, Vi, h, u, uh, ep, cr)

Arguments

par

Current parameters being optimized.

parms

Indices or names of parameters in 'par' to be updated.

pair

A string indicating the pair of variables (e.g., "temperature-wind") being analyzed.

par_all

Complete set of parameters for the model.

data

3D array of observed data, with dimensions corresponding to times, locations, and variables.

names

Vector of variable names, indicating the variables' names (e.g., "temperature" and "wind").

Vi

Matrix where each line corresponds for a possible combination of variables in "names".

h

Vector of spatial distances for the pair.

u

Vector of temporal distances for the pair.

uh

Matrix containing pairs of spatial and temporal distances, and additional information.

ep

A matrix or data frame defining pairs of variables.

cr

Correlation matrix, initial or base correlations between variables.

Value

The log-likelihood value for the given pair of variables based on the current model parameters.

Details

This function implements the methods described in Section 3.3 of the article Stochastic Environmental Research and Risk Assessment, 2025 (DOI: 10.1007/s00477-024-02897-8).