Ordered Normalization

Function for ordered normalization using Ordered Quantile Normalizing transformation.

orderNorm(x, left, n_logit_fit = min(length(x), 1e+05), ..., warn = TRUE)

Arguments

x

Numeric vector to be normalized.

left

Numeric value specifying the left truncation.

n_logit_fit

The number of points to use in the logistic regression for extrapolation, default is the lesser of the length of x or 100,000.

...

Additional arguments passed to the glm function.

warn

Logical flag indicating whether to issue a warning when ties are detected in the data.

Value

A list with class 'orderNorm' containing the normalized data (x.t), the original data (x), and a fitted model object (fit) for future extrapolation.