Plot Observed vs. Simulated Density

Compares the density of observed and simulated data for specific weather variables across seasons. It generates density plots for each variable, distinguishing between observed and simulated data, and arranges them by season. This function relies on `ggplot2` for plotting and `patchwork` for combining plots.

plot_observed_vs_simulated_density(
  sim,
  observed,
  dates,
  seasons,
  location,
  names,
  names_seasons = NULL
)

Arguments

sim

A 3-dimensional array of simulated weather data with dimensions [time, location, variable].

observed

A 3-dimensional array of observed weather data with dimensions [time, location, variable].

dates

A vector of dates corresponding to the first dimension of `sim` and `observed`.

seasons

A list where each element defines the start and end days and months of a season.

location

The index of the location for which to plot the data.

names

A vector of strings indicating the names of the variables to be plotted.

names_seasons

(Optional) A vector of season names corresponding to the `seasons` list. If `NULL`, the names are derived from the `seasons` list itself.

Value

A `ggplot` object representing the combined density plots for observed and simulated data across the specified seasons and variables.

Examples

if (FALSE) {
  # Assuming 'sim', 'observed', 'dates', 'seasons', and 'coordinates' are already defined:
  
  # Define the seasons (example)
  seasons <- list(
    s1 = list(min_day = 1, max_day = 29, min_month = 12, max_month = 2),
    s2 = list(min_day = 1, max_day = 31, min_month = 3, max_month = 5),
    s3 = list(min_day = 1, max_day = 31, min_month = 6, max_month = 8),
    s4 = list(min_day = 1, max_day = 30, min_month = 9, max_month = 11)
  )
  
  # Define variable names to be plotted
  names <- c("Precipitation", "Wind", "Temp_max")
  
  # Specify the location index (for example, 1 for the first location)
  j <- 1
  
  # Run the plot function (this example assumes that the necessary data is loaded)
  plot_result <- plot_observed_vs_simulated_density(sim, observed, dates, seasons, j, names)
  
  # Print or view the plot
  print(plot_result)
}